+++ scripts/action-forge.js
@@ -19,7 +19,7 @@ Hooks.once("init", () => {
     });
   }
 
-  console.info(`PF2E Action Forge | Initialized ${module?.version ?? "0.1.0-dev.4"}`);
+  console.info(`PF2E Action Forge | Initialized ${module?.version ?? "0.1.0-dev.4.2"}`);
 });
 
 Hooks.on("getSceneControlButtons", (controls) => {